% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.r
\name{convert_to_}
\alias{convert_to_}
\title{Convert multiple columns in data frame at once}
\usage{
convert_to_(type, df, columns)
}
\arguments{
\item{type}{A type to convert the columns into. One of "character".
"numeric", "factor", and "logical"}

\item{df}{A data frame.}

\item{columns}{A list of column names to change the type of.}
}
\value{
A data frame.
}
\description{
Helper function to easily convert multiple columns of a data frame into a
speficic type to avoid forcing the type of each column individually.
}
\examples{
example_df <- data.frame(
  condition = c("a", "b", "a", "b"),
  participant = c(1, 1, 2, 2)
)
example_df <- convert_to("factor", example_df, c("condition", "participant"))
}
