% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.r
\name{rgba}
\alias{rgba}
\title{CSS-compliant rgba-definitions}
\usage{
rgba(red, green, blue, alpha = 1)
}
\arguments{
\item{red}{An integer between 0 and 255.}

\item{green}{An integer between 0 and 255.}

\item{blue}{An integer between 0 and 255.}

\item{alpha}{An integer between 0 and 1.}
}
\value{
A hex colour value.
}
\description{
Helper function to easily create colours in accordance with CSS standards
(red, green, and blue specified in intervals from 0 to 255 and alpha
specified in the interval from 0 to 1)
}
\examples{
theme_colour <- rgba(156, 97, 20)
}
